[config] move admission_control_config to own file
per title
