+++ config/src/config.rs
@@ -26,10 +26,11 @@ use std::{
     path::{Path, PathBuf},
     str::FromStr,
     string::ToString,
-    time::Duration,
 };
 use toml;
 
+mod admission_control_config;
+pub use admission_control_config::*;
 mod consensus_config;
 pub use consensus_config::*;
 mod execution_config;
@@ -162,28 +163,6 @@ impl Default for LoggerConfig {
         LoggerConfig {
             is_async: true,
             chan_size: 256,
-        }
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-#[serde(default)]
-pub struct AdmissionControlConfig {
-    pub address: String,
-    pub admission_control_service_port: u16,
-    pub need_to_check_mempool_before_validation: bool,
-    pub max_concurrent_inbound_syncs: usize,
-    pub upstream_proxy_timeout: Duration,
-}
-
-impl Default for AdmissionControlConfig {
-    fn default() -> AdmissionControlConfig {
-        AdmissionControlConfig {
-            address: "0.0.0.0".to_string(),
-            admission_control_service_port: 8000,
-            need_to_check_mempool_before_validation: false,
-            max_concurrent_inbound_syncs: 100,
-            upstream_proxy_timeout: Duration::from_secs(1),
         }
     }
 }
